// src/api/generated/api.ts
export const DomainModelsEnumsEnumAuth = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2
} as const;

export type DomainModelsEnumsEnumAuth = typeof DomainModelsEnumsEnumAuth[keyof typeof DomainModelsEnumsEnumAuth];


export interface ServerControllersModelsAdministratorDTO {
    'id'?: string;
    'surname'?: string | null;
    'name'?: string | null;
    'patronymic'?: string | null;
    'phoneNumber'?: string | null;
    'username'?: string | null;
}
export interface ServerControllersModelsAuthUserDTO {
    'id'?: string;
    'username'?: string | null;
    'passwordHash'?: string | null;
    'role'?: ServerControllersModelsEnumsEnumAuth;
}


export interface ServerControllersModelsClientDTO {
    'id'?: string;
    'companyName'?: string | null;
    'phoneNumber'?: string | null;
}
export interface ServerControllersModelsCreateAdministratorRequestDto {
    'phoneNumber'?: string | null;
    'surname'?: string | null;
    'name'?: string | null;
    'patronymic'?: string | null;
    'username'?: string | null;
    'password'?: string | null;
}
export interface ServerControllersModelsEmployeeDTO {
    'id'?: string;
    'surname'?: string | null;
    'name'?: string | null;
    'patronymic'?: string | null;
    'phoneNumber'?: string | null;
    'task'?: string | null;
    'plantDomain'?: string | null;
    'administratorId'?: string;
}

export const ServerControllersModelsEnumsEnumAuth = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2
} as const;

export type ServerControllersModelsEnumsEnumAuth = typeof ServerControllersModelsEnumsEnumAuth[keyof typeof ServerControllersModelsEnumsEnumAuth];
// ...
export interface ServerControllersModelsUpdateUserRoleRequestDto {
    'newRole'?: DomainModelsEnumsEnumAuth;
}